Support building from references
We try to do a ton of `git` operations inside the docker container while
building docs. Without this change if one of those operations points to
a git repo that was created with `--reference` but wasn't
`--dissociate`d then we'd fail because with didn't mount the reference.
This mounts the reference into the docker container and uses
`GIT_ALTERNATE_OBJECT_DIRECTORIES` to pick them up.